streamlit_forecaster/requirements.txt
yfinance>=0.2.50
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
lightgbm>=4.5
xgboost>=2.1
pmdarima>=2.0
prophet>=1.1
matplotlib>=3.9
statsmodels>=0.14
scipy>=1.11
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.39
pyyaml>=6.0
shap>=0.45
jinja2>=3.1